Fix remote execution in fuzzer (#24)
