+++ crates/wgslsmith/src/harness_runner.rs
@@ -1,0 +1,134 @@
+use std::fmt::{Display, Write as _};
+use std::io::{self, BufRead, BufReader, BufWriter, Write as _};
+use std::path::PathBuf;
+use std::process::{Child, Command, ExitStatus, Stdio};
+use std::thread;
+
+use eyre::eyre;
+use tap::Tap;
+
+#[derive(PartialEq, Eq)]
+pub enum ExecutionResult {
+    Success,
+    Crash(String),
+    Mismatch,
+    // TODO: Detect timeouts from running harness
+    // Might not actually be necessary since it's probably fine to treat them as successful runs
+    // Timeout,
+}
+
+impl Display for ExecutionResult {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ExecutionResult::Success => write!(f, "success"),
+            ExecutionResult::Crash(_) => write!(f, "crash"),
+            ExecutionResult::Mismatch => write!(f, "mismatch"),
+            // ExecutionResult::Timeout => write!(f, "timeout"),
+        }
+    }
+}
+
+pub enum Harness {
+    Local(PathBuf),
+    Remote(String),
+}
+
+pub fn exec_shader(
+    harness: &Harness,
+    config: Option<&str>,
+    shader: &str,
+    metadata: &str,
+    logger: &mut dyn FnMut(String),
+) -> eyre::Result<ExecutionResult> {
+    let mut cmd = match harness {
+        Harness::Local(harness_path) => Command::new(harness_path).tap_mut(|cmd| {
+            cmd.args(["run", "-", metadata]);
+        }),
+        Harness::Remote(remote) => Command::new(std::env::current_exe()?).tap_mut(|cmd| {
+            cmd.args(["remote", remote, "run", "-", metadata]);
+        }),
+    };
+
+    if let Some(config) = config {
+        cmd.args(["-c", config]);
+    }
+
+    let mut harness = cmd
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .stdin(Stdio::piped())
+        .spawn()?;
+
+    {
+        let stdin = harness.stdin.take().unwrap();
+        let mut writer = BufWriter::new(stdin);
+        write!(writer, "{shader}")?;
+        writer.flush()?;
+    }
+
+    let mut stderr = String::new();
+
+    let status = wait_for_child_with_line_logger(harness, &mut |kind, line| {
+        if kind == StdioKind::Stderr {
+            writeln!(stderr, "{line}").unwrap();
+        }
+        logger(line);
+    })?;
+
+    let result = match status.code() {
+        None => return Err(eyre!("failed to get harness exit code")),
+        Some(0) => ExecutionResult::Success,
+        Some(1) => ExecutionResult::Mismatch,
+        Some(101) => ExecutionResult::Crash(stderr),
+        Some(code) => return Err(eyre!("harness exited with unrecognised code `{code}`")),
+    };
+
+    Ok(result)
+}
+
+#[derive(PartialEq, Eq)]
+enum StdioKind {
+    Stdout,
+    Stderr,
+}
+
+fn wait_for_child_with_line_logger(
+    mut child: Child,
+    logger: &mut dyn FnMut(StdioKind, String),
+) -> Result<ExitStatus, io::Error> {
+    let (tx, rx) = crossbeam_channel::unbounded();
+
+    child.stdout.take().map(|stdout| {
+        thread::spawn({
+            let tx = tx.clone();
+            move || {
+                BufReader::new(stdout)
+                    .lines()
+                    .flatten()
+                    .try_for_each(|line| tx.send((StdioKind::Stdout, line)))
+                    .unwrap();
+            }
+        })
+    });
+
+    child.stderr.take().map(|stderr| {
+        thread::spawn({
+            let tx = tx.clone();
+            move || {
+                BufReader::new(stderr)
+                    .lines()
+                    .flatten()
+                    .try_for_each(|line| tx.send((StdioKind::Stderr, line)))
+                    .unwrap();
+            }
+        })
+    });
+
+    drop(tx);
+
+    while let Ok((kind, line)) = rx.recv() {
+        logger(kind, line);
+    }
+
+    child.wait()
+}
